c311: extract TopicsTab cluster (LAMBDA_VALUES + TopicsTab + LdaConfigBadge + DistinguishingWordsGrid + DistinguishingWordsColumn) to own file (#441 2.1)
Workspace.tsx 3420 LOC -> 2732 LOC (-688 LOC, third-largest single extraction).
Entry chunk 93.27 KB -> 70.74 KB (-22.5 KB, biggest single-tab delta in campaign).
ALL 28 visible Workspace tabs now extracted + lazy.
Cumulative since c262: 311 -> 70.74 KB (-77%).
Removed newly-unused IntertopicMap + TopicGraph + TopicSpectrum + TopicSpectrumComparison + TabEmpty imports from Workspace.tsx.
typecheck clean, 44/44 tests pass.
